# src/main/resources/student_report_openapi.yaml
openapi: 3.0.3
info:
  title: Student Report System API
  description: |
    REST API for the Student Report System.
    Built with **Spring Boot + Kotlin**.

    ## Authentication
    All protected endpoints require a **Bearer token** obtained from `POST /auth/login`.
    Include it as: `Authorization: Bearer <token>`

    Tokens are opaque UUIDs validated server-side against the `Sessions` table on
    every request. To invalidate a token immediately (logout, ban), the session row
    is deleted — no wait for expiry required.

    ## Idempotency
    Mutating endpoints (`POST`, `PUT`, `PATCH`) that are marked with the
    `Idempotency-Key` header support safe retries. Supply a unique UUID v4 per
    logical request. Duplicate keys within the TTL window return the cached
    response without re-executing the operation.

    ## Roles
    | Role  | Description                                      |
    |-------|--------------------------------------------------|
    | User  | Student — can submit & manage their own reports  |
    | Admin | Staff  — full access to all resources            |

    ## Status Transitions
    ```
    Pending → In_Progress → Completed
    Pending → Rejected
    In_Progress → Rejected
    ```

  version: 1.2.0
  contact:
    name: Student Report Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development
  - url: https://api.studentreport.example.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Opaque session token issued by POST /auth/login.
        Supply as: `Authorization: Bearer <token>`

  # ─────────────────────────────────────────────────────────────────────────────
  # REUSABLE HEADERS
  # ─────────────────────────────────────────────────────────────────────────────
...
  schemas:
...
    UserRole:
      type: string
      enum: [Admin, User]
      example: User
    ReportStatus:
      type: string
      enum: [Pending, In_Progress, Completed, Rejected]
      example: Pending

    # ── Generic Wrappers ──────────────────────────────────────────────────────
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          nullable: true
    PagedResponse:
      description: |
        Reusable pagination envelope. Inline-extend this with the concrete
        `content` array type at each endpoint using `allOf`.
      type: object
      required: [content, pagination]
      properties:
        content:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          example: 10
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: must be a valid email address

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    UserResponse:
      type: object
      required: [id, name, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Muhammad Guntur Ricky Adhitya
        email:
          type: string
          format: email
          example: 241017310003@mhs.ulm.ac.id
        role:
          $ref: '#/components/schemas/UserRole'
        studentData:
          $ref: '#/components/schemas/StudentDataResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    StudentDataResponse:
      type: object
      required: [id, userId, nim, faculty, major, year]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        nim:
          type: string
          example: "2410173100003"
        faculty:
          type: string
          example: Fakultas Teknik
        major:
          type: string
          example: Teknik Informatika
        year:
          type: integer
          example: 2024
...
    BuildingResponse:
      type: object
      required: [id, name, code, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Gedung Fakultas Teknik Banjarmasin
        code:
          type: string
          example: GFTBJM
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RoomResponse:
      type: object
      required: [id, buildingId, name, floor, code, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        building:
          $ref: '#/components/schemas/BuildingResponse'
        name:
          type: string
          example: Ruang Kuliah A14
        floor:
          type: integer
          minimum: 0
          example: 1
        code:
          type: string
          description: Unique room code across all buildings.
          example: RK-A14
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CategoryResponse:
      type: object
      required: [id, name, description, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Kerusakan Fasilitas
        description:
          type: string
          example: Laporan terkait kerusakan fasilitas kampus
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ReportResponse:
      type: object
      required: [id, reporter, category, room, title, status, upvoteCount, images, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
          description: Optimistic-locking version counter.
          example: 0
        reporter:
          $ref: '#/components/schemas/UserResponse'
        category:
          $ref: '#/components/schemas/CategoryResponse'
        room:
          $ref: '#/components/schemas/RoomResponse'
        title:
          type: string
          example: AC Rusak di Ruang A14
        description:
          type: string
          nullable: true
          example: AC tidak berfungsi sejak kemarin pagi
        status:
          $ref: '#/components/schemas/ReportStatus'
        upvoteCount:
          type: integer
          minimum: 0
          example: 5
        images:
          type: array
          items:
            $ref: '#/components/schemas/ReportImageResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Null when the report has not been soft-deleted.
    CreateReportRequest:
      type: object
      required: [categoryId, roomId, title]
      description: |
        Creates a new report. Initial status is always `Pending` and is set
        server-side; do not include a status field.
      properties:
        categoryId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
          example: AC Rusak di Ruang A14
        description:
          type: string
          nullable: true
          example: AC tidak berfungsi sejak kemarin pagi
    UpdateReportRequest:
      type: object
      minProperties: 1
      description: Only the reporter (while status is Pending) or an Admin may update these fields.
      properties:
        categoryId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
...
    ReportImageResponse:
      type: object
      required: [id, reportId, imageUrl, uploadedAt]
      properties:
        id:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        imageUrl:
          type: string
          format: uri
          example: https://storage.example.com/reports/img_001.jpg
        uploadedAt:
          type: string
          format: date-time

    # ── Upvote ────────────────────────────────────────────────────────────────
...
    ReportLogResponse:
      type: object
      required: [id, reportId, admin, oldStatus, newStatus, createdAt]
      properties:
        id:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        admin:
          $ref: '#/components/schemas/UserResponse'
        oldStatus:
          $ref: '#/components/schemas/ReportStatus'
        newStatus:
          $ref: '#/components/schemas/ReportStatus'
        notes:
          type: string
          nullable: true
          example: Teknisi sudah ditugaskan
        createdAt:
          type: string
          format: date-time

  # ── Common Parameters ──────────────────────────────────────────────────────
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      description: Zero-based page index.
      schema:
        type: integer
        default: 0
        minimum: 0
    QuerySize:
      name: size
      in: query
      description: Number of items per page.
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    QuerySearch:
      name: search
      in: query
      description: Free-text search term (case-insensitive substring match).
      schema:
        type: string
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        UUID v4 for safe retries. If omitted, no idempotency check is applied.
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426614174000

  # ── Common Responses ───────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /reports:
    get:
      tags: [Reports]
      summary: List reports
      description: |
        - **User**: returns only their own non-deleted reports.
        - **Admin**: returns all reports (including soft-deleted) with optional filters.
      operationId: listReports
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QuerySize'
        - $ref: '#/components/parameters/QuerySearch'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReportStatus'
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: roomId
          in: query
          schema:
            type: string
            format: uuid
        - name: buildingId
          in: query
          schema:
            type: string
            format: uuid
        - name: includeDeleted
          in: query
          description: Include soft-deleted reports. Admin only; ignored for Users.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated list of reports
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PagedResponse'
                          - type: object
                            properties:
                              content:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Reports]
      summary: Create a new report (User)
      description: |
        Submits a new report. Status is always initialised to `Pending` server-side.
        Increments the reporter's `UserStats.reportCount` and `UserStats.pendingReport`.
      operationId: createReport
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Referenced category or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /reports/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Reports]
      summary: Get report by ID
      description: |
        - **User**: only accessible if they are the reporter.
        - **Admin**: accessible for any report.
      operationId: getReportById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Reports]
      summary: Update report content (reporter or Admin)
      description: |
        Updates title, description, category, or room. Status changes must use
        `PATCH /reports/{id}/status`.

        Reporters may only edit while the report status is `Pending`.
      operationId: updateReport
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportRequest'
      responses:
        '200':
          description: Updated report
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Reports]
      summary: Soft-delete report (reporter or Admin)
      description: |
        Sets `deleted_at` to the current timestamp. The database row is **not**
        physically removed.

        If the record is later hard-deleted (admin operation), all related
        `Report_Image`, `Upvote`, and `Report_Log` rows are cascade-deleted.
      operationId: deleteReport
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Report soft-deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /reports/{id}/logs:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Report Log]
      summary: Get audit log for a specific report
      description: |
        Returns all status-change log entries for the given report, newest first.
        - **User**: can view logs of any reports.
        - **Admin**: can view and edit logs for any report.

        Log entries are cascade-deleted when the parent report is permanently removed.
      operationId: getReportLogs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QuerySize'
      responses:
        '200':
          description: Paginated log entries for the report
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PagedResponse'
                          - type: object
                            properties:
                              content:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ReportLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /report-logs:
    get:
      tags: [Report Log]
      summary: List all report log entries (Admin only)
      operationId: listAllReportLogs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QuerySize'
        - name: adminId
          in: query
          description: Filter by the admin who made the change.
          schema:
            type: string
            format: uuid
        - name: reportId
          in: query
          description: Filter by report.
          schema:
            type: string
            format: uuid
        - name: newStatus
          in: query
          description: Filter by the resulting status.
          schema:
            $ref: '#/components/schemas/ReportStatus'
      responses:
        '200':
          description: Paginated list of all log entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PagedResponse'
                          - type: object
                            properties:
                              content:
                                type: array
                                items:
                                  $ref: '#/components/schemas/ReportLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /report-logs/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Report Log]
      summary: Get a single log entry by ID (Admin only)
      operationId: getReportLogById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Log entry detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'